+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/java/com/liferay/analytics/reports/web/internal/data/time/TimeRange.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.analytics.reports.web.internal.data.model;
+package com.liferay.analytics.reports.web.internal.data.time;
 
 import java.time.Clock;
 import java.time.LocalDate;
