+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/java/com/liferay/analytics/reports/web/internal/data/time/TimeSpan.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.analytics.reports.web.internal.data.model;
+package com.liferay.analytics.reports.web.internal.data.time;
 
 import com.liferay.portal.kernel.util.Validator;
