+++ modules/dxp/apps/analytics/analytics-reports-web/src/test/java/com/liferay/analytics/reports/web/internal/data/time/TimeRangeTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.analytics.reports.web.internal.data.model;
+package com.liferay.analytics.reports.web.internal.data.time;
 
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
